+++ app/src/main/java/com/example/talapehuser/itmeList.java
@@ -3,41 +3,41 @@ package com.example.talapehuser;
 public class itmeList {
     String name;
     String price;
-    int i;
+    Double i;
     Double total,totaloftotal;
     String date;
 
     private itmeList(){}
 
-    public int getI() {
+    public Double getI() {
         return i;
     }
 
-    public void setI(int i) {
+    public void setI(Double i) {
         this.i = i;
     }
 
-    public itmeList(String name, String price, int i) {
+    public itmeList(String name, String price, Double i) {
         this.name = name;
         this.price = price;
         this.i=i;
 
     }
 
-    public itmeList(String name, String price, int i, Double total) {
+    public itmeList(String name, String price, Double i, Double total) {
         this.name = name;
         this.price = price;
         this.i = i;
         this.total = total;
     }
 
-    public itmeList(String name, String price, int i, String date) {
+    public itmeList(String name, String price, Double i, String date) {
         this.name = name;
         this.price = price;
         this.i = i;
         this.date = date;
     }
-    public itmeList(String name, int i, String date) {
+    public itmeList(String name, Double i, String date) {
         this.name = name;
         this.i = i;
         this.date = date;
